pyproject.toml
[project]
name = "mac-telegrambot-controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pynput>=1.8.1",
    "pytelegrambotapi>=4.30.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
